update failing test
